Fixes to loggedActivityDateCorrect test logic
Small fixes to the test. Also added a new if statement to make the
dates considered equal if the time difference between them is a minute
or less. This is because getting my local time then getting the
websites time causes a delay, and that delay will make inconsistencies
in the time comparison.