Added to old target tests
Added to the sorting target tests:
-basically takes in the number of results after Show More
-then does it's sorting
-takes in the results again
-checks to see if those numbers are the same

Also: So the selectors for the search bar and show more button are
different between the Saved Searches vs Saved Targets. If you want me to
explain in more detail about this topic, hit me up.

*IMPORTANT NOTE* Ok so this has an on-going desktop ticket: desktop
8189, so this test WILL BREAK! But that's ok. Because it's a real issue.
And this test will pass once it's fixed :)

*LESS IMPORTANT NOTE* Also I really hate the name searchSearchInput but
I was following the previous naming-convention, so if anyone has a
better name for that selector let me know.